# src/main/resources/swagger.yaml
swagger: '2.0'
info:
  version: 1.0.0
  title: Dogs in the Vineyard Non-Player Character API
  description: >-
    API used to manage Non-Player Characters (NPCs) in a game of Dogs in the
    Vineyard (DitV)
  contact:
    name: Timo Newton
    email: motipha@gmail.com
schemes:
  - https
produces:
  - application/json
paths:
  '/{game}/batches':
    get:
      tags:
        - NPC Batches
      description: >-
        Retrieve collections of NPC's and Proto-NPC's associated with a specific
        game
      operationId: getBatches
      parameters:
        - name: game
          in: path
          description: unique identifier of the game
          required: true
          type: number
        - name: state
          in: query
          description: 'State of the batch: Active, Unpopulated, Exhausted'
          required: false
          type: string
      responses:
        '200':
          description: list of all batches of NPC's created in this game
          schema:
            type: array
            items:
              $ref: '#/definitions/batch'
        '404':
          description: Unable to find the specified game
        default:
          description: unexpected error
  '/{game}/batch':
    post:
      tags:
        - NPC Batches
      description: Creates a new batch of NPC's
      operationId: createBatch
      parameters:
        - name: game
          in: path
          description: unique identifier of the game
          required: true
          type: number
      responses:
        '201':
          description: NPC Batch has been successfully created
          schema:
            $ref: '#/definitions/batch'
        default:
          description: unexpected error
  '/batch/{batch}':
    parameters:
      - name: batch
        in: path
        description: unique identifier of the batch
        required: true
        type: number
    get:
      tags:
        - NPC Batches
      description: Retrieve a batch of NPC's
      operationId: getBatch
      responses:
        '200':
          description: Requested batch of NPC's
          schema:
            $ref: '#/definitions/batch'
        default:
          description: unexpected error
    put:
      tags:
        - NPC Batches
      description: update a batch of NPC's
      operationId: updateBatch
      consumes:
        - application/json
      parameters:
        - name: batch_payload
          in: body
          description: The NPC batch payload
          required: true
          schema:
            $ref: '#/definitions/batch'
      responses:
        '200':
          description: The NPC Batch
          schema:
            $ref: '#/definitions/batch'
        default:
          description: unexpected error
  '/{game}/npcs':
    get:
      tags:
        - NPCs
      description: retrieves all defined NPC's for this game
      operationId: getNPCs
      parameters:
        - name: game
          in: path
          description: Unique identifier of the game
          required: true
          type: number
      responses:
        '200':
          description: all NPC's in the specified game
          schema:
            type: array
            items:
              $ref: '#/definitions/npc'
        default:
          description: unexpected error
  '/npc/{npc}':
    parameters:
      - name: npc
        in: path
        description: Unique identifier of the npc
        required: true
        type: number
    get:
      tags:
        - NPCs
      description: retrieves a NPC for this game
      operationId: getNPC
      responses:
        '200':
          description: NPC
          schema:
            $ref: '#/definitions/npc'
        default:
          description: unexpected error
    put:
      tags:
        - NPCs
      description: updates a NPC for this game
      operationId: updateNPC
      responses:
        '200':
          description: NPC
          schema:
            $ref: '#/definitions/npc'
        default:
          description: unexpected error
definitions:
  batch:
    type: array
    items:
      $ref: '#/definitions/npc'
  npc:
    properties:
      name:
        type: string
      status:
        type: string
        enum:
          - proto
          - alive
          - dead
      stats:
        type: array
        items:
          $ref: '#/definitions/stat'
      traits:
        type: array
        items:
          $ref: '#/definitions/trait'
      relationships:
        type: array
        items:
          $ref: '#/definitions/trait'
  diePool:
    required:
      - size
      - number
    properties:
      size:
        type: integer
      number:
        type: integer
  stat:
    required:
      - name
      - dice
    properties:
      name:
        type: string
      dice:
        $ref: '#/definitions/diePool'
  trait:
    required:
      - wording
      - dice
    properties:
      wording:
        type: string
      dice:
        type: array
        items:
          $ref: '#/definitions/diePool'